requirements.txt
Flask==2.3.0
mysql-connector-python==9.0.0